Repo: Add a shell script to deal with Windows chmod issues
When working with both WSL and Windows, executable flags are frequently messed up and there appears to be no good (read: easy and reliable) solution.

Since this is the umpteenth time that I'd had to google the specific git incantations to add the flag manually, maybe this will make it slightly less annoying to deal with.