CI: Streamline the existing GitHub Actions workflows
They've always been a bit messy, and to some extent that can't be helped (since they use YAML and that's... not a great language). But leaving in some "temporarily" removed steps is pointless since they can trivially be restored via copy/paste, when I get around to it.

However, I want to trim them down to the simplest possible form and also split Mac/Unix runs so they're easier to maintain.