Tests: Add a missing test case for detailed bdd runs
When the spec file itself contains an error outside of any BDD section,  such as trying to require a module that doesn't exist at the top level, the error will be eaten by the xpcall and not reported.

The error is still recognized and added to the error details table, but since there's no sections to report as failing it will report the remaining sections as passing, leading to strange oddities like "all tests are passing" being displayed and then dumping the error details on the user.

The easiest way to fix this is to just fail loudly, as there's no point in continuing with evaluating sections if the error is critical. I'm assuming here that an error that's not a failed assertion is critical, of course...