Deps: Add a default package source to the WebView2 script
When NuGet is first run on a fresh system, there might not be a local package config to fall back on. For some reason the default source appears to not be used in this case unless explicitly specified.

This problem could potentially be very confusing, so let's just forcibly set the package source (even if that may need updating in the future) to make sure it always works on the first try.