Repo: Ensure that the build tools are in package.path
The path seems to be set correctly in the CI, as well as on my system. But when trying out a "fresh" MSYS2 install on my laptop I noticed LUA_PATH wasn't set - presumably because I haven't installed Lua for Windows or anything else.

I suppose it could be set in the shell scripts directly, but the code should be portable and so this seems a bit easier to manage.