Runtime: Add a new console library to aid testing
It's also useful in some other scenarios, where output needs to be buffered for whatever reason. And it can serve as a base to build proper logging on later on, which is a planned feature anyway.

It's simple enough so there isn't much lost if I end up scrapping it, too.