CI: Add caching via ccache to the Mac OS build workflow
The homebrew dependency variable helps to reduce the build time:

Without it, brew automatically updates all the dependencies of ccache, which include heavyweights such as LLVM. This takes about 5 minutes for downloading several gigabytes of data - which defeats the point of using ccache to speed up the process.

This may cause issues (according to brew devs), so let's wait and see.