CI: Add caching via sccache to the Windows build workflow
Since ccache isn't working properly on Windows (or at least, the action isn't fully supported),  using sccache is a reasonable workaround.

The environment variable it provides is a Windows path and doesn't work in the MSYS2-based OpenSSL build step (which cannot be changed due to the reliance on the Perl distribution from MSYS).