Tests: Add a workaround for testing multiple webviews
This needs to be fixed properly so that we can create multiple webviews and destroy them at will, without SEGFAULT on OSX.

Right now we can't, and I don't particularly feel like digging through the arcane Objective C code in webview itself, so this will have to do. It's tracked via #77 so the info won't get lost.