Runtime: Streamline the displayed OpenSSL version format
All other embedded library versions display only the version number, not the version of the bindings. OpenSSL stands out visually here, drawing undue attention to the line and generally breaking the flow of the version command's output. Maybe it's OCD but I'd prefer the output to be consistent.

Unfortunately, there is no way to read the OPENSSL_VERSION_* constants to derive only the version and I don't want to hijack the bindings code from C++. Maybe that can be upstreamed at some point, but for now just extracting it from the final string should suffice (assuming the format doesn't change).