Repo: Update the LuaJIT build script to patch bcsave on MSYS2
Embedding symbols in the wrong format just causes a warning to be emitted before the info is stripped by the linker, but it's pretty annoying as the output clutters the console window.

Changing the format only affects MSVC, which isn't a supported toolchain on Windows (and likely won't be anytime soon), so let's get it over with. It's also trivially easy to revert if push comes to shove.