Deps: Revert a problematic miniz commit to fix INFLATE
The decompressor returns an empty string in all cases, because of the early exit that the upstream added a while ago. Until that's fixed, the library cannot be used at all and so I manually overrode the change.

Using a git patch should at least help keep the submodule clean.