Runtime: Add a last_error function to the miniz bindings
This is a bit questionable, but without it there's no way to actually test the error messages (since Lua doesn't have access to the archive userdata, and it likely shouldn't). The map is just to guard against multithreaded access even if I don't really consider this supported...

The previous approach of hardcoding errors doesn't work reliably.