Runtime: Add support for WebGPU to the GLFW bindings
This allows the GLFW API to utilize a WebGPU surface as its designated render target, which is currently the only way of displaying anything inside a GLFW window.

In order to retrieve the GLFW window's backing store (surface) in a format that WebGPU can understand (and render to), a third-party extension is currently required because the author's PR hasn't yet been merged by the GLFW maintainers.