Runtime: Fix image decoding not returning RGBA pixel data
This seems like an oversight: Setting the number of channels to 4 has no effect if using stbi_load_image since that'll use the default pixel format (and set the channels to whatever was found in the file).

For some images this doesn't make a difference, but when loading PNG or JPG it definitely can mess things up. The most intuitive approach for me is to always just use RGBA, and if another format is needed either use the FFI bindings directly or (if need be) add specific APIs to return the image data in the desired format.