Runtime: Introduce a new FFI-based crypto library
This complements the existing lua-openssl bindings with an optimized implementation of Base64 (required for Argon2 MCF generation). Benchmarks show that it's about 3-10x times faster when using the FFI (as opposed to the C API), which I consider significant enough to warrant the inclusion in this module.

Needless to say, the surface area of any cryptographic code should be kept minimal, which is why there's only glue code binding to OpenSSL APIs, and a slightly modified version of the PHC Argon2 Base64 implementation.

Key derivation and verification via Argon2 is NYI (coming soon).