Build: Enable -O3 optimizations to speed up Base64 encoding
It seems that the Base64-encoder in the Argon2 library benefits greatly from this. With just -O2 enabled, it takes almost twice as long to decode a randomized set of inputs compared to OpenSSL's EVP-based implementation. With -O3, however, it's actually about equal (after manually applying some basic optimizations, that is).