Runtime: Reorder the parameters of miniz checksum APIs
The documentation has this inverted, presumably because it was the more intuitive order. I understand that checksums are "objects" internally and so passing it as the first parameter might make sense from a technical perspective, but as a user of the API it's confusing.

The most common use case is "give me the checksum of this string", and optimizing for the common case means moving the IV so that it feels more optional than the actual input - even if, technically speaking, both are optional (as per the code).