Deps: Fix various issues in the lua-openssl build scripts
This fixes the CI error that suddenly popped up.

It should also make sure the local OpenSSL version is used instead of the system one, in cases where it isn't. That part is admittedly a bit of a Hail Mary since CMake might still somehow manage to do the wrong thing, but it is what it is.