Runtime: Add a workaround for zlib patch versions being nil
The code assumes that zlib.version always returns three number values, but that's not the case: In zlib 1.3 the patch version isn't set, which is unfortunate but not something that can be changed non-intrusively.

If this becomes a bigger problem then it would be possible to override the lzlib bindings, but for now I'd rather avoid it. Instead, let's just encode the changed expectation and call it a day...