Runtime: Ensure a clean shutdown of the uws event loop
There are a number of handles that uws creates internally. These need to be unhooked from the libuv event loop (and the Lua VM's state) as otherwise the runtime will crash when the Lua state is being closed.

The luv shutdown sequence involves calling all of the now-invalid handles, which results in a SEGFAULT if any of the uws handles is still linked to the event loop (and corresponding Lua state). This is exactly what happens when deleting the LuaVirtualMachine.