Runtime: Add a new etrace library for event tracing
This is effectively (part of) the old C_EventSystem namespace, but easier to use and better tested (IMHO).

There's still some missing features like support for logging levels (requires the planned `syslog` library, which is NYI) or event emitters/handlers, which are currently a bit messy and non-standardized. I'll get to that stuff eventually...