Tests: Add a test case for nesting before and after calls
This should be comprehensive enough to ensure there are no similar surprises if the test runner has to be changed again in the future.

Definitely not the most elegant way of testing this, but the smoke tests are limited to very few primitive tools in order to ensure the runtime can always be bootstrapped from scratch.