Runtime: Implement a WebGPU render interface for RML
It's async in the sense of deferred event queues, even if access to the file system (loading fonts, documents, and texture files) is currently blocking.

This render interface doesn't stand on its own, unlike the builtin ones that ship with RML. Instead, the idea is to keep the bare minimum of state and logic in the C++ layer as its more or less opaque to LuaJIT.

Because that's in line with the GLFW bindings, the code makes use of the same interoperability layer. While there's some performance overhead for passing on the various events, it doesn't seem reasonable to pointlessly copy the generated geometry or textures, so the glue layer creates GPU resources that can then be managed by LuaJIT (e.g., via finalizers).

Needless to say, this design is still somewhat complicated by the fact that two environments need to be juggled and RML only exposes callbacks. Perhaps in the future this could be simplified by a tighter integration or by moving more of the graphics context into the runtime?