Tests: Increase the delay for various HTTP integration tests
These are somewhat flaky because the entire approach to testing is poorly thought-out. I don't have time to implement a better one right now (and it doesn't matter until there's a reason to change the code), so for now just work around the flakiness by reducing the chance of messages not being sent in time. Of course the tests will take longer, but their execution time currently pales in comparison to the build time, so that's irrelevant.