Runtime: Add a shorthand for starting main.lua to the CLI
SInce this is the designated entry point for evo-based apps, and there's no way to override it directly, having a convenient shorthand for this use case seems handy. This is equivalent to a typical "start" command (like npm start) but without the awkwardness of changing the default mode of operation in a way that works differently from the CLI of PUC Lua and LuaJIT, and without the need for the overhead of an added command.