CI: Disable build workflow runs on feature branch creation
Not sure where I copy/pasted that part of the configuration from, but it's definitely triggering the build workflows to run when branches are first created (which is pointless and wasteful). Before a PR is opened and merged, there's really no point in running any CI workflows, especially with how long they take to complete.