Build: Enable the use of C sources via new build rules
This isn't as clean as I'd have liked, but it's the easiest way to build lua-utf8 (since it won't compile with g++). Thankfully, the build generator is extensible enough to deal with all sorts of inputs, even if there's now a little more duplication that might need to be cleaned up later.