Runtime: Make the CLI output a little more colorful
This should help with readability, at least it does on my terminal.

I also took the liberty to reorganize the output functions a bit since that made it a lot easier to colorize. The color scheme itself is limited to a few select colors, used more or less consistently, but could use some more streamlining (later).

Some of the phrasing is updated to improve the consistency of CLI interactions, though the meaning is hopefully unchanged.