Runtime: Improve the failure mode for CLI commands
Previously, the CLI always exited with EXIT_SUCCESS even if the command displayed the usage info (a failure mode that should probably result in EXIT_FAILURE). This is unfortunate because it means other programs can't detect that a command, e.g. a build or test runner, has failed. The snapshot test suite has been updated to receive more inputs, but the `tmpfile` thing is a bit sketchy.