Runtime: Fix a bug in the default oop constructor
Derived classes using the default constructor should be able to access the prototype's properties if they don't override them. However, this is only possible if the instance is created from the base class (inheriting its properties), and not an empty table (that has no access to properties of the parent).