Runtime: Fix the order of args being passed to zip apps
Normally, scripts are run from the interpreter CLI. In that case, arg[0] is the script name - just like when using PUC Lua or LuaJIT itself. However, when building a LUAZIP app with the build command, and subsequently running it with the same args as before, arg[0] is now the first actual CLI arg because there's no interpreter CLI that would have to receive the script name.

If shifting everything by one, the script should be able to function the same way regardless of whether it's running from the interpreter CLI or as a standalone app. Storing the app name instead of the script name is a bit questionable, and therefore untested (=undefined behavior).

fmt