Runtime: Reverse the order of displayed error details
Sometimes there can be a cascade of errors, e.g,. because an earlier test failed and leaves the environment in an inconsistent state. In such scenarios, seeing tons of errors in the console is to be expected. However, the natural approach of trying to fix the error that's immediately in view first (i.e., the last one displayed in the console) will only lead to frustration as there might not be anything wrong with the test case itself.

In GCC and other tools there's a way to stop cascading errors and halt execution after the first, which however doesn't seem desirable here. Instead, an arguably more useful approach: The errors can simply be reported in reverse order. This way, the last one displayed is actually the first and no scrolling will be needed, but no information is lost.