Runtime: Create FFI bindings for the LuaJIT REPL
It might seem overkill to set up a new FFI binding just for this, but I've been wanting to streamline the interpreter CLI and some of the setup code that uses globals and defines, as well as the build configuration, anyway. This library can  be used as a container for those "random" things only really required by the runtime itself, while still providing a proper API instead of hacky global access.