Runtime: Enable the LuaJIT REPL via the eval command
The implementation currently just does nothing, which is arguably not very helpful. It could display a help text, but there's really little point. The command itself was mostly added so that debuggers in IDEs like VS Code work (since both LuaJIT and the PUC Lua interpreter provide an eval command that is used by debuggers).

Just getting no output might even be confusing, so starting the REPL seems like a good default choice here that also adds the need for adding a separate REPL command (bloats the CLI even more). Lastly, using the REPL has the same use cases for manually debugging or quickly checking something as the eval command (but it's easier to use as no shell escape nonsense will interfere with the process).