Runtime:  Fix event notifications for disabled events
The behavior as specified doesn't make any sense, as it conflates the event log with the actual event registry. Disabling logging should of course not disable notifications as that would completely break any code that expects to get notified.

Maybe this is a sign the intent needs to be made clearer, seeing how the "minimal" naming convention selected here (which is typically used in Lua and C) doesn't sufficiently express this nuance. Will have to consider this when streamlining the API...