Refactor: Create a separate library for the FFI bindings
It's probably easier to streamline the startup process if the entry point is created ahead-of-time and populated with the FFI bindings. If this is done from Lua, as is currently the case, then the load order of the core libraries and the FFI bindings can interfere with each other, causing minor annoyances and errors.

It also is a bit awkward having to pass the exports table as a magic global, especially if they aren't ever accessed again that way once they're initialized and exported to their respective packages.