Runtime: Streamline access to the FFI export tables
In order to make accessing static exports less verbose, can automatically create aliases based on the list of exported functions. This requires some jumping through some hoops because the cdata can't be iterated over like a regular table.

Of course this isn't a replacement for a more usable high-level API, but it is much less annoying to call those functions if their names are shorter. Either way, the bindings have to be moved over gradually since there are so many - hence the fallbacks.