Runtime: Update the bindings to store embedded cdefs
It would be much easier to maintain the various bindings if the definition weren't duplicated, as keeping them in sync can be error-prone and they're overly verbose in at least some cases.

As a first step in that direction, introduce a distinction between the stored export table and the C definitions (that are currently still stored in both the C++ and Lua sources, hence the placeholder).