Refactor: Move all shared cdefs to separate header files
Unfortunately the types used by LuaJIT's FFI won't always be quite identical as the C++ layer can use the full range of types, whereas LuaJIT often has to make do with opaque pointers for library data structures or callback types.

In order to deduplicate the shared type definitions anyway, they'll have to be stored separately (and without any of the clutter that LuaJIT might be unable to process). This way, the compiler can check everything before LuaJIT even sees them. And most importantly they don't have to be kept in sync manually, which is tedious and error-prone.

Also, library types still need to be copy/pasted - which isn't ideal...