Refactor: Move aliased types to separate header files
The benefit here isn't as great since the compiler can't check the type definitions that are used by LuaJIT only. Still, moving them to separate headers reduces the boilerplate and file size in the Lua sources and enables syntax highlighting in the IDE.

It also makes the dump output more useful for those libraries since it's not includingall the cdefs, but not sure how useful that really is.