Refactor: Store aliased FFI cdefs in separate C headers
These are just the "LuaJIT-only" portion of the existing cdefs, which aren't used by the C++ code (and unfortunately also not type-checked by the compiler). Storing them as C header enables syntax highlighting at least, and in the future they could be compiled on their own just to check that there are no errors?

 Either way, for now they'll be inserted by the sync-cdefs script. This is much easier if they're stored in separate files, so they shall be.