Runtime: Add support for __tostring to json.encode
This is a non-reversible operation, but it's assumed that if someone willingly serializes non-primitive objects and assigns metamethods to them as well, they want to actually serialize the object and not face an error when encoding. This assumption may have to be revisited in the future, but it works with event payloads.