CI: Add a workaround for outdated bit_width signatures
It seems that the return type was changed retroactively, which isn't reflected in the GCC and clang versions installed on GitHub's runners? Either way, just dereferencing the function itself won't compile in that environment.

It might be sensible to upgrade the installed toolchain in the CI runner, but that'll increase build times even more.  It's prudent to test that the build proceeds with "older" versions of the supported toolchains anyway, as well as the relatively recent ones used by MSYS on Windows (which seems to have no issue here).