Runtime: Add a non-standard package.open extension
This is more universally useful as a shorthand during test runs etc., compared to adding a separate export function to the assertions library, or global aliases for bdd primitives.

But mostly I just really want to use a function called package.open() -  it's like opening a gift, except it pollutes the (global) environment!