Runtime: Add an async file reader to the FileSystem API
This module makes heavy use of the event system to flatten the hierarchy of nested callbacks and make things easier to test/observe. Not sure how much I like the design, but it's arguably less convoluted than playing ping-pong with coroutines and should be relatively easy to use, as well. Additional features, like gathering multiple files, can be added on top of this later.