Runtime: Add a utility for resolving dynamic library names
When using ffi.load this is seemingly handled by the underlying platform, but when attempting to load libraries from the VFS that isn't going to work.

There's no way around resolving names manually, since the files have to be extracted to a temporary directory and then loaded from disk (using the fully-qualified file name). Providing a separate function for the lookup  makes it easier to test and also allows users to check that resolution works as expected.