Tests: Add a missing nil test case for validateString
Not that it matters much, but I did run into a silly bug recently that had me question my sanity for a bit...

I then created this test case, only to find out I swapped a parameter, so the validation always worked. Can't hurt to keep the test, though, given that not having one left open the possibility of a bug in the validation - however unlikely.