Build: Strip GDB debug symbols from generated object files
I don't know why this is here, but it doesn't make much sense. If a SEGFAULT in the runtime layer needs to be debugged, developers can add it manually. And if there's a crash in LuaJIT itself, then the symbols aren't going to help anyway. Setting the -g flag for bcsave should be sufficient and takes up much less space in comparison, while enabling readable stack traces.

The GDB symbols increase the binary size by a whopping 10 MB, which certainly doesn't seem desirable. Time for a diet!