Runtime: Fix the typing of the WebServer's uws app handle
Taking a quick look at the examples, I don't think this type was supposed to be used directly. In fact, there's an SSLApp, CachingApp, and the regular App - which means the WebServer was relying on what seems like an implementation detail?

The networking code needs a thorough review anyway, so for now just fix the build error and pretend the other problems don't exist.