Deps: Drop the experimental support for n2 builds
n2 is missing some features and doesn't seem as actively maintained as ninja itself. Sadly, the speedup from "better incremental builds" isn't very noticeable in practice. On the other hand, using n2 instead of ninja in the CI workflow adds a full minute for building Rust code with `cargo`.

I'm not too happy about the Rust toolchain dependency in general (although it's also required for wgpu-native currently), so that's another reason to stick with ninja for the time being. It can still be used with an alias if needed, but I don't know that it's worth supporting two mostly identical build tools when I'd actually prefer to have zero.