Build: Fix some bcsave-related warnings on macOS
Currently the runtime is compiled in Objective C compatibility mode because of glfw3webgpu, which unfortunately still hasn't been merged upstream yet (see 43dbf1e for context). LuaJIT bytecode objects are passed as C files to the linker, but the runtime is of course a C++ executable - hence the warning.

This was previously changed in 8e0af2e when adding support for M1 builds, because the generated bytecode objects were misaligned. I don't know if the alignment issues are still present, but there's only one way to find out...