CI: Add OpenSSL headers to the build cache
The headers are required when building lua-openssl (and libcurl) against the cached OpenSSL libraries. Right now they're not included, which leads CMake to silently fall back to the system-installed version.

While that might be convenient, it's really not desirable here. In fact, it's exactly what I'm trying to avoid. The system version could be much older than the embedded one. Although that's of course unlikely, since distro maintainers will want to keep OpenSSL as updated as possible. In that scenario, building libcurl (and lua-openssl) against different versions could cause all sorts of issues.