Runtime: Export the low-level iconv API as well
The bindings currently only export a decently-fast but rather inflexible iconv wrapper. For some use cases (e.g., gradually converting incoming bytes or resuming failed conversions) it would be useful to have access to all interfaces.

This low-level API is rather difficult to use and also pretty slow, so maybe it'd be better to create an async/continuation-style API in C++ and export that. However, the maintenance cost of just exposing the existing function is arguably lower and it's a niche use case anyway.