Runtime: Enable LUAZIP apps to load shared objects
Relying on a shared library that's created by a build script removes some complexity from the test suite, at the cost of depending on this specific library always being built.

While certainly not ideal, in the case of zlib it's probably safe to assume that this will always be the case.